requirements.txt
appdirs==1.4.3
certifi==2018.11.29
chardet==3.0.4
idna==2.8
packaging==16.8
PyNamecheap==0.0.3
pyparsing==2.2.0
pysqlite==2.8.3
requests==2.21.0
six==1.10.0
urllib3==1.24.1
